chore(api): move country usecase to shared
