+++ api/tests/shared/integration/domain/usecases/find-countries_test.js
@@ -1,7 +1,7 @@
-import { findCountries } from '../../../../../../src/certification/enrolment/domain/usecases/find-countries.js';
-import { domainBuilder, expect, sinon } from '../../../../../test-helper.js';
+import { findCountries } from '../../../../../src/shared/domain/usecases/find-countries.js';
+import { domainBuilder, expect, sinon } from '../../../../test-helper.js';
 
-describe('Certification | Enrolment | Unit | UseCase | find-country', function () {
+describe('Unit | UseCase | find-country', function () {
   let countryRepository;
 
   beforeEach(function () {
